Create synth_button_matrix_test.ino
Sort of works, when pressing buttons in col1, a: strings don't change note (they do if you press them while the string is playing?) b: one of the buttons reads as row1 for col2 as well. no idea why. it does it if you switch the wires to a different pin too. it's not the diodes. maybe it's the mcu pin