Update synth_button_matrix_test.ino
fixed changing not working when not pressed, some issue with the for each loop, didn't like it for some reason.  changed it to a dumb for loop and now it works properly
fixed button reading weird, it was either electrical noise or some capacitance/the diode not being able to charge fast enough. added a small delay before the first read of the next column and the issue is fixed.
still need to fix the "losing notes" bug when you switch root while you're playing a chord, comments are in there